Fix type casting in COM_ParseExt function for array length checks
- Updated the length comparison in the COM_ParseExt function to explicitly cast ARRAY_LEN to an integer, ensuring proper type handling and preventing potential issues with array bounds.

+++ src/qcommon/q_shared.c
@@ -683,7 +683,7 @@ const char *COM_ParseExt( const char **data_p, qboolean allowLineBreaks )
 			{
 				com_lines++;
 			}
-			if ( len < ARRAY_LEN( com_token )-1 )
+			if ( len < (int)ARRAY_LEN( com_token )-1 )
 			{
 				com_token[ len ] = c;
 				len++;
@@ -694,7 +694,7 @@ const char *COM_ParseExt( const char **data_p, qboolean allowLineBreaks )
 	// parse a regular word
 	do
 	{
-		if ( len < ARRAY_LEN( com_token )-1 )
+		if ( len < (int)ARRAY_LEN( com_token )-1 )
 		{
 			com_token[ len ] = c;
 			len++;
